requirements.txt
# RAG 시스템 핵심 라이브러리
langchain-openai>=0.0.2
langchain-google-genai>=0.0.2
langchain-huggingface>=0.0.2
langchain-chroma>=0.1.0
langchain-text-splitters>=0.1.0

# 데이터 처리
pandas>=2.1.0
openpyxl>=3.1.5
tabulate>=0.9.0

# PDF 처리
pdfplumber>=0.10.0 # 테이블 추출에 유리
PyMuPDF4LLM>=0.0.1
llama-parse>=0.0.1
nest-asyncio>=1.0.0
marker-pdf>=0.0.1

# 벡터 데이터베이스
chromadb>=0.4.0

# LLM
openai>=1.10.0
google-genai>=0.2.0

# 웹 프레임워크
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=23.0.0

# 웹 크롤링
beautifulsoup4>=4.12.0
requests>=2.31.0

# 유틸리티
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.26.0
